Refactor code to support multi_algs
