+++ alg_random.py
@@ -1,0 +1,9 @@
+import random
+
+
+# function to simulate play 
+def random_alg(map):
+		movetypes = ("w", "a", "s", "d")
+
+		move = random.choice(movetypes)
+		return move
